requirements.txt
Dropbox==1.5.1
hurry.filesize==0.9

